Relationships are now defined by a hash to be consistent with models
This also enhances the Dbee::Model#hash specs to cover the graph based case.
