+++ lib/dbee/model.rb
@@ -10,13 +10,15 @@
 require_relative 'model/constraints'
 require_relative 'model/partitioner'
 require_relative 'model/relationships'
+require_relative 'util/make_keyed_by'
 
 module Dbee
   # In DB terms, a Model is usually a table, but it does not have to be.  You can also re-model
   # your DB schema using Dbee::Models.
   class Model
-    extend Forwardable
     acts_as_hashable
+    extend Dbee::Util::MakeKeyedBy
+    extend Forwardable
 
     class ModelNotFoundError < StandardError; end
 
@@ -26,34 +28,6 @@ module Dbee
     def_delegator :models,          :sort,    :sorted_models
     def_delegator :constraints,     :sort,    :sorted_constraints
     def_delegator :partitioners,    :sort,    :sorted_partitioners
-
-    class << self
-      # Given a hash of hashes or a hash of values of instances of this class,
-      # a hash is returned where all of the values are instances of this class.
-      # An ArgumentError is raised if the value's <tt>key_attrib</tt> is not
-      # equal to the top level hash key.
-      def make_keyed_by(key_attrib, spec_hash)
-        # Once Ruby 2.5 support is dropped, this can just use the block form of
-        # #to_h.
-        spec_hash.map do |key, spec|
-          [key, make_value_checking_key_attib!(key_attrib, key, spec)]
-        end.to_h
-      end
-
-      private
-
-      def make_value_checking_key_attib!(key_attrib, key, spec)
-        if spec.is_a?(self)
-          if spec.send(key_attrib).to_s != key.to_s
-            err_msg = "expected a #{key_attrib} of '#{key}' but got '#{spec.send(key_attrib)}'"
-            raise ArgumentError, err_msg
-          end
-          spec
-        else
-          make((spec || {}).merge(key_attrib => key))
-        end
-      end
-    end
 
     def initialize(
       name:,
@@ -65,8 +39,7 @@ module Dbee
     )
       @name           = name
       @constraints    = Constraints.array(constraints || []).uniq
-      # TODO: raise an error if two relationships share a name
-      @relationships  = Relationships.array(relationships || []).to_set
+      @relationships  = Relationships.make_keyed_by(:name, relationships)
       @models_by_name = name_hash(Model.array(models))
       @partitioners   = Partitioner.array(partitioners).uniq
       @table          = table.to_s.empty? ? @name : table.to_s
@@ -100,7 +73,7 @@ module Dbee
     end
 
     def relationship_for_name(relationship_name)
-      relationships.find { |relationship| relationship.name == relationship_name }
+      relationships[relationship_name]
     end
 
     def ==(other)
