+++ lib/dbee/schema_from_tree_based_model.rb
@@ -25,7 +25,7 @@ module Dbee
           name: tree_model.name,
           table: tree_model.table,
           partitioners: tree_model.partitioners,
-          relationships: []
+          relationships: {}
         }
 
         parent_model && add_relationship_to_parent_model(
@@ -40,7 +40,7 @@ module Dbee
       end
 
       def add_relationship_to_parent_model(model, graph_model_attrs)
-        graph_model_attrs[:relationships].push(name: model.name, constraints: model.constraints)
+        graph_model_attrs[:relationships][model.name] = { constraints: model.constraints }
       end
     end
   end
